requirements.txt
django
django-user-media
django-hvad
Pillow
django-libs
django-multiselectfield